Move some functionality to groupby/numba_.py
